Fixes after clang code model analysis. Explicitly type casting, copy constructors and others, floating point operations optimized. Transition to new signal-slot syntax (in progress), C++14 enabled now.

+++ atlastranslator.h
@@ -19,11 +19,11 @@ public:
 private:
     QSslSocket sock;
     QString atlHost;
-    int atlPort;
+    quint16 atlPort;
     bool inited;
     bool emptyRestore;
 public:
-    explicit CAtlasTranslator(QObject *parent, QString host, int port, const QString& SrcLang);
+    explicit CAtlasTranslator(QObject *parent, QString host, quint16 port, const QString& SrcLang);
     virtual ~CAtlasTranslator();
 
     bool initTran();
